requirements.txt
click>=8.1.3
requests>=2.31.0
python-dotenv>=1.0.0
packaging>=23.0
halo>=0.0.31
pyperclip>=1.8.2
netifaces>=0.11.0
speedtest-cli>=2.1.3
psutil>=5.9.0
cryptography>=42.0.2
Pillow>=10.0.0
pydub>=0.25.1
moviepy>=1.0.3
